coroutines refacto, new entity, cleanup model, update schema, fix Intent
* coroutines refacto : still not sure of current state of things.
However each flow seems now to run in the right Dispatcher. Default for
long running task, Main for UI. It seems a bit hackish but responsive
* new entity: I need to save alarm data to be able to reactivate them
when relaunching the app and set the screen right when returning on the
station tile. I've added a new entity for that
* cleanup model : citistationinformationmodel had fields that were not
used by my code. It's a bit silly because if Citibike removes them from
the api, my app crashes so I should set only what's necessary
* Intent: retrieving Intent is a funny business and I've decided to
declare the station Id in the action of the intent. Consequently, I will
have 2 potential pending intent for one station id. One for the alarm,
one for the geofencing